Tailor Matcher implementation more to the tokenization process
Removes some general-purpose DAFSA functions from named_character_references.zig and makes Matcher the primary public API, while also adding some more state to Matcher in order to make implementing HTML tokenization simpler.

Matcher also now calculates the unique index as it walks the DAFSA and stores the unique index of the last match, which means:

- No longer need to store the matched string in order to look up the associated codepoints
- No longer need to rewalk the DAFSA in order to calculate the unique index

This simplifies the implementation in test.zig and improves the performance of the DAFSA when benchmarking against HTML with tens of thousands of named character references.