Reduce size of second layer by using a contiguous array of 'second layer' nodes
Instead of a struct-of-arrays for each first character, the second layer now uses a single contiguous array of second layer nodes and the 'first to second layer link' stores an offset into that array. Each node uses 4 bytes (whereas before it was 3), but this still reduces the overall size of the data because we remove the pointer fields from the 'struct of arrays' second layer struct.

Has ~no performance difference